% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{detect_cuda}
\alias{detect_cuda}
\title{Detect cuda}
\usage{
detect_cuda()
}
\value{
boolean, whether cuda is available.
}
\description{
This function detects whether cuda is available. If \code{setup_grafzahl} was executed with \code{cuda} being \code{FALSE}, this function will return \code{FALSE}. Even if \code{setup_grafzahl} was executed with \code{cuda} being \code{TRUE} but with any factor that can't enable cuda (e.g. no Nvidia GPU, the environment was incorrectly created), this function will also return \code{FALSE}.
}
