% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_amharic_data}
\alias{get_amharic_data}
\title{Download The Amharic News Text Classification Dataset}
\usage{
get_amharic_data()
}
\value{
A named list of two corpora: training and test
}
\description{
This function downloads the training and test sets of the Amharic News Text Classification Dataset from Hugging Face.
}
\references{
Azime, Israel Abebe, and Nebil Mohammed (2021). "An Amharic News Text classification Dataset." arXiv preprint arXiv:2103.05639
}
