% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{predict.grafzahl}
\alias{predict.grafzahl}
\title{Prediction from a fine-tuned grafzahl object}
\usage{
\method{predict}{grafzahl}(object, newdata, cuda = detect_cuda(), return_raw = FALSE, ...)
}
\arguments{
\item{object}{an S3 object trained with \code{\link[=grafzahl]{grafzahl()}}}

\item{newdata}{a \link[quanteda:corpus]{quanteda::corpus} or a character vector of texts on which prediction should be made.}

\item{cuda}{logical, whether to use CUDA, default to \code{\link[=detect_cuda]{detect_cuda()}}.}

\item{return_raw}{logical, if \code{TRUE}, return a matrix of logits; a vector of class prediction otherwise}

\item{...}{not used}
}
\value{
a vector of class prediction or a matrix of logits
}
\description{
Make prediction from a fine-tuned grafzahl object.
}
