% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{supported_model_types}
\alias{supported_model_types}
\title{Supported model types}
\format{
An object of class \code{character} of length 23.
}
\usage{
supported_model_types
}
\description{
A vector of all supported model types.
}
\keyword{datasets}
