% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{unciviltweets}
\alias{unciviltweets}
\title{A Corpus Of Tweets With Incivility Labels}
\format{
An object of class \code{corpus} (inherits from \code{character}) of length 19982.
}
\usage{
unciviltweets
}
\description{
This is a dataset from the paper "The Dynamics of Political Incivility on Twitter". The tweets were by Members of Congress elected to the 115th Congress (2017–2018). It is important to note that not all the incivility labels were coded by human. Majority of the labels were coded by the Google Perspective API. All mentions were removed. The dataset is available from Pablo Barbera's Github. \url{https://github.com/pablobarbera/incivility-sage-open}
}
\references{
Theocharis, Y., Barberá, P., Fazekas, Z., & Popa, S. A. (2020). The dynamics of political incivility on Twitter. Sage Open, 10(2), 2158244020919447.
}
\keyword{datasets}
