% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{detect_conda}
\alias{detect_conda}
\alias{detect_cuda}
\title{Detecting Miniconda And Cuda}
\usage{
detect_conda()

detect_cuda()
}
\value{
boolean, whether the system is available.
}
\description{
These functions detects miniconda and cuda.
}
\details{
\code{detect_conda} conducts a test to check whether 1) a miniconda installation and 2) the grafzahl miniconda environment exist.

\code{detect_cuda} checks whether cuda is available. If \code{setup_grafzahl} was executed with \code{cuda} being \code{FALSE}, this function will return \code{FALSE}. Even if \code{setup_grafzahl} was executed with \code{cuda} being \code{TRUE} but with any factor that can't enable cuda (e.g. no Nvidia GPU, the environment was incorrectly created), this function will also return \code{FALSE}.
}
