% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{hydrate}
\alias{hydrate}
\title{Create a grafzahl S3 object from the output_dir}
\usage{
hydrate(output_dir, regression = FALSE)
}
\arguments{
\item{output_dir}{string, location of the output model. Important: Please note that if this directory exists, it will be overwritten.}

\item{regression}{logical, if \code{TRUE}, the task is regression, classification otherwise.}
}
\description{
Create a grafzahl S3 object from the output_dir
}
